add likes, fix permissions to users

+++ djangogramm_app/views.py
@@ -35,62 +35,7 @@ class PostCreateView(View):
             }
             return render(request, self.template_name, context)
         else:
-            messages.success(request, ('You Must Be Loged In To View This Page'))
-            return redirect('login')
-
-    def post(self, request: HttpRequest) -> HttpResponse:
-        if request.user.is_authenticated:
-            picture_form = PictureFormPost(request.POST, request.FILES)
-            tag_form = TagForm(request.POST)
-            post_form = PostForm(request.POST or None, request.FILES or None)
-            images = request.FILES.getlist('picture')
-            if post_form.is_valid() and tag_form.is_valid() and picture_form.is_valid():
-                post = post_form.save(commit=False)
-                post.user = request.user
-                post.save()
-                tags_names = tag_form.cleaned_data['tags'].split(',')
-                tags = []
-                imgs = []
-                for tag_name in tags_names:
-                    tag, created = Tag.objects.get_or_create(name=tag_name.strip())
-                    tags.append(tag)
-                for img in images:
-                    image = Pictures(picture=img)
-                    image.save()
-                    imgs.append(image)
-                post.tags.set(tags)
-                post.pictures.set(imgs)
-                post_form.save_m2m()
-                return redirect('index')
-        else:
-            post_form = PostForm(request.POST, request.FILES, instance=request.user.id)
-            tag_form = TagForm(request.POST)
-            picture_form = PictureFormPost(request.POST, request.FILES)
-        context = {
-            'post_form': post_form,
-            'tag_form': tag_form,
-            'picture_form': picture_form
-        }
-        return render(request, self.template_name, context)
-
-
-class UpdatePostView(View):
-    template_name = 'posts/post_update.html'
-
-    def get(self, request: HttpRequest, post_id: int) -> HttpResponse:
-        if request.user.is_authenticated:
-            current_post = Post.objects.get(id=post_id)
-            picture_form = PictureFormPost()
-            post_form = PostForm(instance=current_post)
-            tag_form = TagForm()
-            context = {
-                'post_form': post_form,
-                'tag_form': tag_form,
-                'picture_form': picture_form
-            }
-            return render(request, self.template_name, context)
-        else:
-            messages.success(request, ('You Must Be Loged In To View This Page'))
+            messages.success(request, ('You Must Be Loged In To Create New Post'))
             return redirect('login')
 
     def post(self, request: HttpRequest) -> HttpResponse:
@@ -133,25 +78,25 @@ class DeletePostView(View):
     template_name = 'posts/post_delete.html'
 
     def get(self, request: HttpRequest, post_id: int) -> HttpResponse:
-        if request.user.is_authenticated:
-            post = Post.objects.get(id=post_id)
+        post = Post.objects.get(id=post_id)
+        if request.user.is_authenticated and post.user == request.user:
             context = {
                 'post': post
             }
             return render(request, self.template_name, context)
         else:
-            messages.success(request, ('You Must Be Loged In To View This Page'))
-            return redirect('login')
+            messages.success(request, ('You Cannot Delete Other Posts!'))
+            return redirect('index')
 
     def post(self, request: HttpRequest, post_id: int) -> HttpResponse:
-        if request.user.is_authenticated:
-            post = Post.objects.get(id=post_id)
+        post = Post.objects.get(id=post_id)
+        if request.user.is_authenticated and post.user == request.user:
             post.delete()
-            messages.success(request, ('Your Post Has Been Deleted'))
+            messages.success(request, ('Your Post Has Been Deleted!'))
         else:
-            messages.success(request, ('You Must Be Loged In To View This Page'))
+            messages.success(request, ('You Must Be Loged In To View This Page!'))
             return redirect('login')
-        return redirect('index')
+        return redirect('profile', request.user.id)
 
 
 class LikePostView(View):
@@ -167,5 +112,22 @@ class LikePostView(View):
                 like_post.likes.add(user)
             return redirect('index')
         else:
-            messages.success(request, ('You Must Be Loged In To View This Page'))
+            messages.success(request, ('You Must Be Loged In To View This Page!'))
+            return redirect('login')
+
+
+class DisLikePostView(View):
+
+    def post(self, request: HttpRequest, post_id: int) -> HttpResponse:
+        if request.user.is_authenticated:
+            dislike_post = get_object_or_404(Post, id=post_id)
+            user = request.user
+            if user in dislike_post.dislikes.all():
+                dislike_post.dislikes.remove(user)
+                return redirect('index')
+            else:
+                dislike_post.dislikes.add(user)
+            return redirect('index')
+        else:
+            messages.success(request, ('You Must Be Loged In To View This Page!'))
             return redirect('login')
